app refactor
