+++ backend/config/fakeKeys.js
@@ -1,3 +1,9 @@
-const keys = {google: {id: 'fake!', secret: 'shh!'}};
+const keys = {
+  google: {
+    id: 'fake!',
+    secret: 'shh!'
+  }
+  // more auth strategies to follow?
+};
 
 module.export = keys;
