+++ backend/routes/authRoutes.js
@@ -1,0 +1,14 @@
+const passport = require('passport');
+
+module.exports = (app) => {
+  app.get('/auth/google', passport.authenticate('google', {scope: ['profile', 'email']}));
+
+  app.get(
+    '/auth/google/callback',
+    passport.authenticate('google', {failureRedirect: '/login'}),
+    function (req, res) {
+      // Successful authentication, redirect home.
+      res.redirect('/');
+    }
+  );
+};
